Splitting out the slice module
