+++ src/matrix/decomposition/bidiagonal.rs
@@ -102,8 +102,7 @@ impl<T> Matrix<T>
 
 #[cfg(test)]
 mod tests {
-    use matrix::Matrix;
-    use matrix::slice::BaseMatrix;
+    use matrix::{BaseMatrix, Matrix};
 
     fn validate_bidiag(mat: &Matrix<f64>,
                        b: &Matrix<f64>,
