+++ src/matrix/impl_ops.rs
@@ -1,6 +1,6 @@
 use super::{Matrix, MatrixSlice, MatrixSliceMut};
 use super::{Row, RowMut, Column, ColumnMut};
-use super::slice::{BaseMatrix, BaseMatrixMut};
+use super::{BaseMatrix, BaseMatrixMut};
 
 use super::super::utils;
 use super::super::vector::Vector;
